Checkpointing

Flux already provides some methods for checkpointing; however, because of its agnostic nature with respect to the data-saving backend, the naming conventions and the choice of objects that need saving, it calls for writing quite an extensive boilerplate code. IMUDevNNTrainingLib makes all those choices for the user and in return streamlines the checkpointing process. In particular, the following choices are made:

JLD2 is used as a backend for saving the state of the trained Neural Net

A checkpoint comprises of saved:

model_state: parameters of the trained Neural Network

opt_state: parameters of the optimizer

log: a training log (with loss functions)

other: a dictionary of any other parameters that the user wishes to save. For instance, plateau_detector may be saved here.

Names of checkpoints take the format checkpoint=$(epoch).jld2, where $(epoch) is the index of the epoch at the end of which the checkpoint has been made.

The main object is Checkpointer:

IMUDevNNTrainingLib.Checkpointer — Type

Checkpointer(;
    dir::String = ".",
    continue_from::Union{Int,Symbol} = :last,
    save_every::Int = 1)

A struct to manage checkpoints during training of neural networks.

dir::String: Root directory where the checkpoints will be saved to

continue_from::Union{Int64, Symbol}: A default behaviour for resuming checkpointing. See start! for details.

save_every::Int64: The frequency of saving checkpoints. A checkpoint will be saved every save_every epochs.

Examples

Checkpointing during training

# a function that updates the training log
model = ...
update_log!(...) = ...
ch = Checkpointer(dir=joinpath(homedir(), "checkpoints"),
                  continue_from=:last,
                  save_every=5)
model, opt_state, log, start_epoch, other = start!(ch, model)
for epoch in start_epoch:100
    Flux.train!(model, data, opt_state)
    update_log!(log, model, data, epoch)
    checkpoint(ch, epoch; model, opt_state, log)
end

Loading checkpoint for testing

model = Chain(...)
ch = Checkpointer(dir=joinpath(homedir(), "checkpoints"))
load!(model, path_to_checkpoint(ch, index_of_last_checkpoint(ch)))
test(model, data)

The usual training workflow involves start!ing it (which will load the specified checkpoint) as well as checkpointing it on every epoch, which will save the model when appropriate:

IMUDevNNTrainingLib.start! — Function

start!(cp::Checkpointer, model;
       move_to_device=Flux.cpu,
       fallback_optimizer=m -> Flux.setup(Flux.Adam(1e-3), m),
       continue_from=cp.continue_from)

Start the checkpointing process. continue_from can be one of the following:

:none, :start, :restart, :nothing: for starting from scratch;

:last, :latest, :recent: for continuing from the last checkpoint;

Int: for continuing from a specific checkpoint index.

The move_to_device function is used to move the model to the desired device (e.g. CPU or GPU). The fallback_optimizer function is used to setup the optimizer if the checkpoint does not exist. The function returns the model, the optimizer state, the training log, the index of the subsequent training epoch and the dictionary of other objects that were saved in the checkpoint.

IMUDevNNTrainingLib.checkpoint — Function

checkpoint(cp::Checkpointer, epoch::Int; model, opt_state, log, kwargs...)

Save the model, the optimizer state and the training log to a checkpoint file if the given epoch is a multiple of cp.save_every. Otherwise, do nothing.

Note

Sometimes the user may wish to save and, perhaps, update additional variables. For instance a common use case would be to save the state of the PlateauDetector. In that case the following could be done:

model = ...
ch = Checkpointer()
model, opt_state, log, start_epoch, other = start!(ch, model)
pd = other[:plateau_detector]

and if we'd like to start using a different schedule of learning rates we could adjust them with:

new_schedule = ParameterSchedulers.Stateful(Exp(1e-3, 0.15))
Flux.adjust!(opt_state, pd, new_schedule)

before we resume the training.

For testing it is often enough to simply load! the model and skip all the other objects:

IMUDevNNTrainingLib.load! — Function

load!(model, checkpoint_path)

Load the model, the optimizer statem, the training log and the other variables from the checkpoint file.

To pick the appropriate checkpoint one can make use of the following helper functions:

IMUDevNNTrainingLib.path_to_checkpoint — Function

path_to_checkpoint(cp::Checkpointer, epoch::Int)

Return the path to the checkpoint file for the given epoch.

Warning

This function does not check if the file exists.

IMUDevNNTrainingLib.index_of_last_checkpoint — Function

index_of_last_checkpoint(cp::Checkpointer)

Return the index of the most recent checkpoint in the directory cp.dir. Return nothing if no checkpoint exists.

IMUDevNNTrainingLib.index_of_last_checkpoint_prior_to — Function

index_of_last_checkpoint_prior_to(cp::Checkpointer, i::Int)

Return the index of the most recent checkpoint in the directory cp.dir out of checkpoints that are prior to the given index i. Return nothing if no such checkpoint exists.

A typical example of loading for testing is given below:

# define model architecture
model = ...

chkp = Checkpointer(; dir="saved_checkpoints")
path = path_to_checkpoint(chkp, index_of_last_checkpoint(chkp))
IMUDevNNTrainingLib.load!(model, path)

# move to GPUs if needed
if USE_GPUs
    model = model |> gpu
end

To list available checkpoint indices you may use:

IMUDevNNTrainingLib.list_existing_checkpoint_indices — Function

list_existing_checkpoint_indices(cp::Checkpointer)

List the epoch indices of the existing checkpoints in the directory cp.dir.