Plateau detector

It is often recommended to reduce the learning rate when the tracked loss metric has stopped improving for some period of time (for instance, in Python, various Neural Net libraries implement ReduceLROnPlateau, see keras's or pytorch's implementations). To this end we introduce a PlateauDetector that tracks the target loss metric and updates learning rate in case it stops improving for a sufficient number of ticks.

Note

We will use the word tick to refer to a sequence of operations that results in PlateauDetector checking the loss metric precisely once. It is up to the user to decide how often that happens.

Note

Unlike python's ReduceLROnPlateau we don't restrict the reduction of the learning rate to be of the form of an exponential decay. Instead, we let the user choose the schedule along which the learning rate is being modified using the interface of Stateful defined in ParameterSchedulers.jl. In particular, an exponential decay could be defined via:

using IMUDevNNTrainingLib
using ParameterSchedulers
using ParameterSchedulers: Stateful

init_learning_rate = 1e-3
decay = 0.7
scheduler = Stateful(Exp(init_learning_rate, decay))
pd = PlateauDetector(; scheduler)

IMUDevNNTrainingLib.PlateauDetector — Type

PlateauDetector(; patience=10, ϵ=1e-8,
                  scheduler=ParameterSchedulers.Stateful(Exp(0.001, 0.7))

A plateau detector for schedulers of learning rate. It will adjust the learning rate when the loss has stopped improving for a number of ticks.

Tip

Most commonly one tick is equal either to one epoch or one batch, but the user is free to define it as they see fit.

patience::Int64: The number of consecutive ticks with no improvement before advancing the learning rate along the scheduler.

last_tick::Int64: Internal only. The last tick number on which the learning rate was adjusted.

best_loss::Float64: Internal only. The best loss found so far.

best_tick::Int64: Internal only. The tick number of the best loss found so far.

ϵ::Float64: The minimum learning rate. If the learning rate is reduced to a value smaller than this, it will be clipped at this value.

scheduler::ParameterSchedulers.Stateful: The schedule along which the learning rate will be adjusted.

IMUDevNNTrainingLib.step! — Function

step!(pd::PlateauDetector, loss::Real)

Advance the plateau detector by one tick. Update pd internals with information about the current loss. Return true if the learning rate should be updated.

Optimisers.adjust! — Function

Flux.Optimisers.adjust!(o, pd::PlateauDetector)

Update the learning rate of the optimizer o according to the schedule defined by the plateau detector pd.

Flux.Optimisers.adjust!(o, pd::PlateauDetector, i::Int)

Reset the learning rate by moving to state i along the schedule defined by pd.scheduler.

Flux.Optimisers.adjust!(o, pd::PlateauDetector, s::ParameterSchedulers.Stateful)

Reset the learning rate by changing the scheduler to s and moving to the state defined by it.

Tip

These last two three-parameter functions are sometimes convenient when we want to manually restart the learning rate schedule (for instance, after loading a checkpoint and restarting training).

Warning

To avoid name clashes step! is not exported by IMUDevNNTrainingLib. It is recommended to use it as follows:

using IMUDevNNTrainingLib
using Flux
const NNTrLib = IMUDevNNTrainingLib
pd = PlateauDetector(; η=1e-4)
optimizer_state = ...
loss = ...
needs_update = NNTrLib.step!(pd, loss)
if needs_update
    Flux.adjust!(optimizer_state, pd)
end

Additionally the following utility function is implemented:

IMUDevNNTrainingLib.learning_rate — Function

learning_rate(pd::PlateauDetector)

Return the current learning rate of the plateau detector pd.

An overloaded Flux.adjust! makes it possible to change the previously set schedules and adjust the state of the optimizer accordingly.

See also the examples/plateau_detector.jl file for more details on how to use a PlateauDetector.