Pretty printing

IMUDevNNTrainingLib implements a handful of useful pretty printing features. To accompany them there are a couple of routines that extract dimension/size information from the MLUtils.DataLoader holding data of the type TemporalData. These are:

IMUDevNNLib.num_samples — Function

IMUDevNNLib.num_samples(loader::DataLoader{<:TemporalData})

The number of availble samples held by the loader (i.e. the maximal batch size).

IMUDevNNTrainingLib.feature_dim — Function

feature_dim(loader::DataLoader{<:TemporalData})

The dimension of the feature space of the data held by the loader. For TemporalData the feature dimension is simply the size of the state vector.

Note

If the state is a multidimensional tensor, feature_dim is equal to the total number of elements in a state tensor.

IMUDevNNTrainingLib.target_dim — Function

target_dim(loader::DataLoader{<:TemporalData})

The dimension of the target space of the data held by the loader. For TemporalData the target dimension is simply the size of the observation vector.

Note

If the observation is a multidimensional tensor, target_dim is equal to the total number of elements in an observation tensor.

IMUDevNNTrainingLib.batch_size — Function

batch_size(loader::DataLoader)

The batch size of the loader.

All the above, as well as some additional ones, can be extracted all at once from the loader either in a form of a NamedTuple or a formatted String using convenience methods:

IMUDevNNTrainingLib.basic_info — Function

basic_info(loader::DataLoader)

Return a NamedTuple with basic dimensional information about the loader and the data it holds.

IMUDevNNTrainingLib.basic_info_as_string — Function

basic_info_as_string(loader::DataLoader{<:TemporalData})

Return a formatted string with basic dimensional information about the loader and the data it holds.

However, all of the above are rarely used directly. Instead, they are being called by the method start_info (intended to be called at the beginning of training) that nicely formats the results:

IMUDevNNTrainingLib.start_info — Function

start_info(loader::DataLoader)

Print basic information about the loader and the data it holds.

Tip

This method is intended to be called at the start of training.

Additionally, the following could be called during training to display progress:

IMUDevNNTrainingLib.echo_epoch — Function

echo_epoch(id)

Print the current epoch number.

IMUDevNNTrainingLib.echo_summary — Function

echo_summary(; epoch, avg_loss, elapsed, title="Training")

Print a summary of the training process.