Checkpointing

Lux is agnostic with respect to the way the user checkpoints the trained models. In particular the data-saving backend, the naming conventions and the choice of objects that need saving are all left up to the user. Although great from the perspective of flexibility, it calls for writing quite an extensive boilerplate code. IMUDevNNTrainingLib makes all those choices for the user and in return streamlines the checkpointing process. In particular, the following choices are made:

JLD2 is used as a backend for saving the state of the trained Neural Net

A checkpoint comprises of saved:

model_parameters: parameters of the trained Neural Network

model_states: states (i.e. non-trainable parameters) of the trained Neural Network

opt_state: parameters of the optimizer

log: a training log (with loss functions)

other: a dictionary of any other parameters that the user wishes to save. For instance, plateau_detector may be saved here.

Names of checkpoints take the format checkpoint=$(epoch).jld2, where $(epoch) is the index of the epoch at the end of which the checkpoint has been made.

The main object is Checkpointer:

IMUDevNNTrainingLib.Checkpointer — Type

Checkpointer(;
    dir::String = ".",
    continue_from::Union{Int,Symbol} = :last,
    save_every::Int = 1)

A struct to manage checkpoints during training of neural networks.

Note

The checkpointer is intended to save:

the model parameters

the model state

the optimizer state

the training log

other objects passed by the user

Note in particular that it is not recommended to save the model itself.

dir::String: Root directory where the checkpoints will be saved to

continue_from::Union{Int64, Symbol}: A default behaviour for resuming checkpointing. See start for details.

save_every::Int64: The frequency of saving checkpoints. A checkpoint will be saved every save_every epochs.

Examples

Checkpointing during training

using Lux, Optimisers
using Random

rng = Random.default_rng()
Random.seed!(rng, 0)

# a function that updates the training log
model = ...
update_log!(...) = ...
ch = Checkpointer(dir=joinpath(homedir(), "checkpoints"),
                  continue_from=:last,
                  save_every=5)
chkp_data, start_epoch = start(ch)
parameters, states, opt_state, log, other = chkp_data

for epoch in start_epoch:100
    train!(parameters, states, model, data, opt_state)
    update_log!(log, model, data, epoch)
    checkpoint(ch, epoch; parameters, states, opt_state, log, other)
end

Loading checkpoint for testing

model = Chain(...)
ch = Checkpointer(dir=joinpath(homedir(), "checkpoints"))
chkp_data = load_checkpoint(model, path_to_last_checkpoint(ch))
test(model, chkp_data, data)

The usual training workflow involves starting it (which will load the specified checkpoint) as well as checkpointing it on every epoch, which will save the model when appropriate:

IMUDevNNTrainingLib.start — Function

start(cp::Checkpointer;
      move_to_device=Lux.cpu_device(),
      continue_from=cp.continue_from)

Start the checkpointing process. continue_from can be one of the following:

:none, :start, :restart, :nothing: for starting from scratch;

:last, :latest, :recent: for continuing from the last checkpoint;

Int: for continuing from a specific checkpoint index.

The move_to_device function is used to move the model parameters, model states and optimizer's state to the desired device (e.g. CPU or GPU). The function returns a NamedTuple with the checkpointed data (or nothing if no checkpoint is found) and an index of the subsequent training epoch.

IMUDevNNTrainingLib.checkpoint — Function

checkpoint(cp::Checkpointer, epoch::Int; model, opt_state, log, kwargs...)

Save the model, the optimizer state and the training log to a checkpoint file if the given epoch is a multiple of cp.save_every. Otherwise, do nothing.

Note

Sometimes the user may wish to save and, perhaps, update additional variables. For instance a common use case would be to save the state of the PlateauDetector. In that case the following could be done:

model = ...
ch = Checkpointer()
chkp_data, start_epoch = start(ch)
if isnothing(chkp_data)
    do_custom_initialization()
end
pd = chkp_data.other[:plateau_detector]

and if we'd like to start using a different schedule of learning rates we could adjust them with:

new_schedule = ParameterSchedulers.Stateful(Exp(1e-3, 0.15))
Optimisers.adjust!(chkp_data.opt_state, pd, new_schedule)

before we resume the training.

For testing it is often enough to call load_checkpoint directly, instead of trying to establish the starting epoch as well:

IMUDevNNTrainingLib.load_checkpoint — Function

load_checkpoint(path::String; move_to_device=Lux.cpu_device())

Load the model parameters, model states, optimizer state, training log and the remaining variables from the checkpoint file. move_to_device is used to move the:

model parameters

model states and

optimizer's state

to the desired device (e.g. CPU or GPU).

To pick the appropriate checkpoint one can make use of the following helper functions:

IMUDevNNTrainingLib.path_to_checkpoint — Function

path_to_checkpoint(cp::Checkpointer, epoch::Int)

Return the path to the checkpoint file for the given epoch.

Warning

This function does not check if the file exists.

IMUDevNNTrainingLib.index_of_last_checkpoint — Function

index_of_last_checkpoint(cp::Checkpointer)

Return the index of the most recent checkpoint in the directory cp.dir. Return nothing if no checkpoint exists.

IMUDevNNTrainingLib.path_to_last_checkpoint — Function

path_to_last_checkpoint(cp::Checkpointer)

Return the path to the most recent checkpoint file for the given epoch.

IMUDevNNTrainingLib.index_of_last_checkpoint_prior_to — Function

index_of_last_checkpoint_prior_to(cp::Checkpointer, i::Int)

Return the index of the most recent checkpoint in the directory cp.dir out of checkpoints that are prior to the given index i. Return nothing if no such checkpoint exists.

A typical example of loading for testing is given below:

# define model architecture
model = ...

chkp = Checkpointer(; dir="saved_checkpoints")
chkp_data = load_checkpoint(model, path_to_last_checkpoint(chkp))

To list available checkpoint indices you may use:

IMUDevNNTrainingLib.list_existing_checkpoint_indices — Function

list_existing_checkpoint_indices(cp::Checkpointer)

List the epoch indices of the existing checkpoints in the directory cp.dir.